# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: مصنع نبع صافيا - نظام إدارة المخزن والمبيعات
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: قائمة المنتجات
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: إضافة منتج
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  # ===================== CUSTOMERS =====================
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: قائمة العملاء
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: إضافة عميل
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/coupon-books:
    get:
      operationId: listCustomerCouponBooks
      tags: [customers]
      summary: دفاتر كوبونات العميل
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CouponBook"
    post:
      operationId: createCouponBook
      tags: [customers]
      summary: إضافة دفتر كوبونات للعميل
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponBookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponBook"

  # ===================== REPS =====================
...
  /inventory/transactions:
    get:
      operationId: listInventoryTransactions
      tags: [inventory]
      summary: حركات المخزن
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [in, out]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"

  # ===================== EXPENSES =====================
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: قائمة المصروفات
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: إضافة مصروف
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    patch:
      operationId: updateExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, color, unitPrice, vatRate]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
          enum: [white, blue]
        unitPrice:
          type: number
        vatRate:
          type: number
        description:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, color, unitPrice, vatRate]
      properties:
        name:
          type: string
        color:
          type: string
          enum: [white, blue]
        unitPrice:
          type: number
        vatRate:
          type: number
        description:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        unitPrice:
          type: number
        vatRate:
          type: number
        description:
          type: string

    # ---- CUSTOMER ----
    Customer:
      type: object
      required: [id, name, phone, area, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        totalCouponBooks:
          type: integer
        remainingCoupons:
          type: number
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        notes:
          type: string

    # ---- COUPON BOOK ----
    CouponBook:
      type: object
      required: [id, customerId, totalValue, remainingValue, bookNumber, issuedAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        bookNumber:
          type: string
        totalValue:
          type: number
        remainingValue:
          type: number
        issuedAt:
          type: string
        notes:
          type: ["string", "null"]
    CouponBookInput:
      type: object
      required: [totalValue, bookNumber]
      properties:
        bookNumber:
          type: string
        totalValue:
          type: number
        notes:
          type: string

    # ---- REP ----
...
    InventoryTransaction:
      type: object
      required: [id, date, type, productId, productName, quantity, reference, referenceType, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
          enum: [in, out]
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        reference:
          type: string
        referenceType:
          type: string
          enum: [purchase, sale]
        createdAt:
          type: string

    # ---- EXPENSE ----
    Expense:
      type: object
      required: [id, date, category, description, amount, createdAt]
      properties:
        id:
          type: integer
        date:
          type: string
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [date, category, description, amount]
      properties:
        date:
          type: string
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        notes:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        date:
          type: string
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        notes:
          type: string

    # ---- TREASURY ----
...