# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * مصنع نبع صافيا - نظام إدارة المخزن والمبيعات
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseUpdate {
  date?: string;
  category?: string;
  description?: string;
  amount?: number;
  notes?: string;
}